Move Nautilus workspace members to Rust Stable
Nautilus is one of the main features keeping LibAFL on the nightly
channel.

The features of `exclusive_range_pattern` and `step_trait` are both
nonessentially used in the code base and seem to not be being merged
into rust stable anytime soon.

By trivially removing the test the repo compiles with Rust stable with
an MSRV of 1.58 (Released 2022-01-13).

I tested LibAFL with the "nautilus" feature pointed to this commit and
the feature remains unchanged.